+++ packages/components/src/components/ContentfulContainer/ContentfulContainerAsHyperlink.tsx
@@ -6,6 +6,7 @@ import type {
   ExperienceDataSource,
   ExperienceUnboundValues,
   StyleProps,
+  DragWrapperProps,
 } from '@contentful/experiences-core/types';
 
 import { EntityStore } from '@contentful/experiences-core';
@@ -30,7 +31,7 @@ export type ContentfulContainerAsHyperlinkProps<EditorMode = boolean> = (EditorM
   cfHyperlink?: StyleProps['cfHyperlink'];
   cfOpenInNewTab?: StyleProps['cfOpenInNewTab'];
   WrapperComponent?: React.ElementType;
-  dragProps?: any; // TODO: add type
+  dragProps?: DragWrapperProps;
 };
 
 export const ContentfulContainerAsHyperlink: React.FC<ContentfulContainerAsHyperlinkProps> = (
