+++ packages/core/src/types.ts
@@ -382,3 +382,50 @@ export type BackgroundImageOptions = {
   quality?: string;
   targetSize: string;
 };
+
+interface DraggableProvidedDraggableProps {
+  'data-rfd-draggable-context-id'?: string;
+  'data-rfd-draggable-id'?: string;
+}
+
+interface DraggableProvidedDragHandleProps {
+  'data-rfd-drag-handle-draggable-id'?: string;
+  'data-rfd-drag-handle-context-id'?: string;
+}
+
+export type WrapperTags = keyof Pick<
+  JSX.IntrinsicElements,
+  | 'div'
+  | 'span'
+  | 'section'
+  | 'article'
+  | 'aside'
+  | 'p'
+  | 'h1'
+  | 'h2'
+  | 'h3'
+  | 'h4'
+  | 'h5'
+  | 'h6'
+  | 'header'
+  | 'footer'
+  | 'nav'
+  | 'main'
+>;
+
+export interface DragWrapperProps
+  extends DraggableProvidedDragHandleProps,
+    DraggableProvidedDraggableProps,
+    React.HTMLAttributes<HTMLElement>,
+    React.PropsWithChildren {
+  'data-cf-node-id'?: string;
+  'data-ctfl-draggable-id'?: string;
+  'data-test-id'?: string;
+  'data-cf-node-block-id'?: string;
+  'data-cf-node-block-type'?: string;
+  'data-ctfl-dragging-element'?: string;
+  innerRef?: (refNode: HTMLElement) => void;
+  wrapComponent?: boolean;
+  Tag?: WrapperTags;
+  ToolTipAndPlaceholder?: React.ReactNode;
+}
